requirements.txt
altair==6.0.0
annotated-types==0.7.0
anyio==4.12.0
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
faiss-cpu==1.13.1
filelock==3.20.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
mpmath==1.3.0
narwhals==2.13.0
networkx==3.6.1
numpy==2.3.5
openai==2.6.1
packaging==25.0
pandas==2.3.3
pillow==12.0.0
protobuf==6.33.2
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
PyMuPDF==1.26.7
PyPDF2==3.0.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.2.1
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.3
sentence-transformers==5.2.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.52.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
torch==2.9.1
tornado==6.5.3
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.2
watchdog==6.0.0
xlsxwriter==3.2.9
